requirements.txt
Django>=5.2.7
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
Pillow>=10.0
django-tailwind>=3.7.0
 xhtml2pdf>=0.2.10
django-crispy-forms>=2.0
django-import-export>=3.3.0
